Split webpack configuration into its compile and dev server counter parts
There is no unique conifguration for the dev server so there does not exist a file for it. It is just using the "core webpack configuration" for this project. The compile script does have a few unique configurations. Difference to note: the webpack-bundle-analyzer does not work in dev server mode (where it outputs a static file). We essentially do not care about it anyways.