Remove `lodash.find` in favor of native solution
Lodash find method is general purpose and about 64KB. This is a ridiculous dependency for simple array manipulation. There exists Array#find but it does not have IE support. Was not sure if a tiny polyfill was necessary. Might make sense to set babel-polyfills as a requirement for this library. That is another effort though.